Android: scale the output bitmap to the view size
The original base64-encoded bitmap tended to be very large for hi-res
cameras; this caused long loading times loading the base64 string into
an image (data: uri).

To mitigate the output bitmap is now scaled to the original view size,
and the JPEG compression is down to 85 from 95.

Additionally, I removed the newlines from the base64 output.